Fixed minor issues related to module config
